Log queued RPC response failures
Add structured sequencer admin response failure logging and cast queued RPC timeout millisecond fields to u64 for tracing subscribers.

+++ crates/consensus/service/src/actors/rpc/engine_rpc_client.rs
@@ -82,7 +82,7 @@ impl QueuedEngineRpcClient {
                 error!(
                     target: "block_engine",
                     response_name,
-                    timeout_ms = timeout.as_millis(),
+                    timeout_ms = timeout.as_millis() as u64,
                     "Timed out waiting for engine rpc response"
                 );
                 ErrorObject::owned(
